Predio

+++ src/model/Quarto.java
@@ -29,7 +29,7 @@ public class Quarto extends ModelDefault{
 	
 	@NotNull(message="Informe o valor da diária!")
 	@Column(name="vlrQua", length=12)
-	private int valorQuarto;
+	private Double valorQuarto;
 	
 	@NotNull(message="Informe o status do quarto!")
 	@Column(name="staQua", length=2)
@@ -82,11 +82,11 @@ public class Quarto extends ModelDefault{
 		this.dormitorios = dormitorios;
 	}
 
-	public int getValorQuarto() {
+	public Double getValorQuarto() {
 		return valorQuarto;
 	}
 
-	public void setValorQuarto(int valorQuarto) {
+	public void setValorQuarto(Double valorQuarto) {
 		this.valorQuarto = valorQuarto;
 	}
